+++ ui/DesignSystem/Primitive/Icon/CurrencyEUR.svelte
