+++ ui/DesignSystem/Primitive/Icon/CurrencyRAD.svelte
