+++ ui/DesignSystem/Primitive/Icon/CurrencyRADSmall.svelte
