+++ ui/DesignSystem/Primitive/Icon/CurrencyUSD.svelte
