+++ ui/DesignSystem/Primitive/Icon/index.js
@@ -24,10 +24,10 @@ import CopySmall from "./CopySmall.svelte";
 import CrossCircle from "./CrossCircle.svelte";
 import Cross from "./Cross.svelte";
 import CrossSmall from "./CrossSmall.svelte";
-import CurrencyDollar from "./CurrencyDollar.svelte";
-import CurrencyEuro from "./CurrencyEuro.svelte";
-import CurrencyRad from "./CurrencyRad.svelte";
-import CurrencyRadSmall from "./CurrencyRadSmall.svelte";
+import CurrencyUSD from "./CurrencyUSD.svelte";
+import CurrencyEUR from "./CurrencyEUR.svelte";
+import CurrencyRAD from "./CurrencyRAD.svelte";
+import CurrencyRADSmall from "./CurrencyRADSmall.svelte";
 import Ellipsis from "./Ellipsis.svelte";
 import EllipsisSmall from "./EllipsisSmall.svelte";
 import Envelope from "./Envelope.svelte";
@@ -93,10 +93,10 @@ export default {
   CrossCircle,
   Cross,
   CrossSmall,
-  CurrencyDollar,
-  CurrencyEuro,
-  CurrencyRad,
-  CurrencyRadSmall,
+  CurrencyUSD,
+  CurrencyEUR,
+  CurrencyRAD,
+  CurrencyRADSmall,
   Ellipsis,
   EllipsisSmall,
   Envelope,
